+++ domain/errors.go
@@ -147,6 +147,32 @@ func (e ConcentratedTickModelNotSetError) Error() string {
 	return fmt.Sprintf("tick model is not set on pool (%d)", e.PoolId)
 }
 
+type AlloyTransmuterDataMissingError struct {
+	PoolId uint64
+}
+
+func (e AlloyTransmuterDataMissingError) Error() string {
+	return fmt.Sprintf("Alloy Transmuter data is missing for pool (%d)", e.PoolId)
+}
+
+type MissingNormalizationFactorError struct {
+	PoolId uint64
+	Denom  string
+}
+
+func (e MissingNormalizationFactorError) Error() string {
+	return fmt.Sprintf("Missing normalization factor for denom (%s) in pool (%d)", e.Denom, e.PoolId)
+}
+
+type ZeroNormalizationFactorError struct {
+	PoolId uint64
+	Denom  string
+}
+
+func (e ZeroNormalizationFactorError) Error() string {
+	return fmt.Sprintf("Normalization factor is zero for denom (%s) in pool (%d)", e.Denom, e.PoolId)
+}
+
 type TransmuterInsufficientBalanceError struct {
 	Denom         string
 	BalanceAmount string