+++ domain/ingester.go
@@ -17,9 +17,11 @@ type GRPCIngesterConfig struct {
 // BlockPoolMetadata contains the metadata about unique pools
 // and denoms modified in a block.
 type BlockPoolMetadata struct {
-	// DenomLiquidityMap is a map of denoms to their liquidities across pools.
+	// DenomPoolLiquidityMap is a map of denoms to their liquidities across pools.
 	// These are constructed from the pool IDs updated within a block.
-	DenomLiquidityMap DenomLiquidityMap
+	DenomPoolLiquidityMap DenomPoolLiquidityMap
+	// UpdatedDenoms are the denoms updated within a block.
+	UpdatedDenoms map[string]struct{}
 	// PoolIDs are the IDs of all pools updated within a block.
 	PoolIDs map[uint64]struct{}
 }