+++ router/delivery/http/router_handler.go
@@ -6,6 +6,8 @@ import (
 	"strconv"
 
 	"github.com/labstack/echo/v4"
+	"go.opentelemetry.io/otel/attribute"
+	"go.opentelemetry.io/otel/trace"
 
 	sdk "github.com/cosmos/cosmos-sdk/types"
 
@@ -64,6 +66,17 @@ func NewRouterHandler(e *echo.Echo, us mvc.RouterUsecase, tu mvc.TokensUsecase, 
 func (a *RouterHandler) GetOptimalQuote(c echo.Context) (err error) {
 	ctx := c.Request().Context()
 
+	span := trace.SpanFromContext(ctx)
+	defer func() {
+		if err != nil {
+			span.RecordError(err)
+			// nolint:errcheck // ignore error
+			c.JSON(domain.GetStatusCode(err), domain.ResponseError{Message: err.Error()})
+		}
+
+		// Note: we do not end the span here as it is ended in the middleware.
+	}()
+
 	isSingleRoute, err := domain.ParseBooleanQueryParam(c, "singleRoute")
 	if err != nil {
 		return c.JSON(domain.GetStatusCode(err), domain.ResponseError{Message: err.Error()})
@@ -86,12 +99,12 @@ func (a *RouterHandler) GetOptimalQuote(c echo.Context) (err error) {
 
 	tokenOutDenom, tokenIn, err := getValidRoutingParameters(c)
 	if err != nil {
-		return c.JSON(domain.GetStatusCode(err), domain.ResponseError{Message: err.Error()})
+		return err
 	}
 
 	chainDenoms, err := mvc.ValidateChainDenomsQueryParam(c, a.TUsecase, []string{tokenIn.Denom, tokenOutDenom})
 	if err != nil {
-		return c.JSON(domain.GetStatusCode(err), domain.ResponseError{Message: err.Error()})
+		return err
 	}
 
 	// Update coins token in denom it case it was translated from human to chain.
@@ -115,7 +128,7 @@ func (a *RouterHandler) GetOptimalQuote(c echo.Context) (err error) {
 
 	quote, err := a.RUsecase.GetOptimalQuote(ctx, tokenIn, tokenOutDenom, routerOpts...)
 	if err != nil {
-		return c.JSON(domain.GetStatusCode(err), domain.ResponseError{Message: err.Error()})
+		return err
 	}
 
 	scalingFactor := oneDec
@@ -125,8 +138,11 @@ func (a *RouterHandler) GetOptimalQuote(c echo.Context) (err error) {
 
 	_, _, err = quote.PrepareResult(ctx, scalingFactor)
 	if err != nil {
-		return c.JSON(domain.GetStatusCode(err), domain.ResponseError{Message: err.Error()})
+		return err
 	}
+
+	span.SetAttributes(attribute.Stringer("token_out", quote.GetAmountOut()))
+	span.SetAttributes(attribute.Stringer("price_impact", quote.GetPriceImpact()))
 
 	return c.JSON(http.StatusOK, quote)
 }