+++ tokens/usecase/pricing/worker/pricing_worker_test.go
@@ -57,21 +57,20 @@ func (s *PricingWorkerTestSuite) TestUpdatePricesAsync() {
 		{
 			name: "empty base denoms",
 			baseDenoms: domain.BlockPoolMetadata{
-				DenomLiquidityMap: domain.DenomLiquidityMap{},
 			},
 		},
 		{
 			name: "one base denom",
 			baseDenoms: domain.BlockPoolMetadata{
-				DenomLiquidityMap: domain.DenomLiquidityMap{UOSMO: {
-					TotalLiquidity: osmomath.OneInt(),
-				}},
+				UpdatedDenoms: map[string]struct{}{
+					UOSMO: {},
+				},
 			},
 		},
 		{
 			name: "several base denoms",
 			baseDenoms: domain.BlockPoolMetadata{
-				DenomLiquidityMap: domain.DenomLiquidityMap{
+				UpdatedDenoms: map[string]struct{}{
 					UOSMO: {},
 					ATOM:  {},
 					USDC:  {},
@@ -81,7 +80,7 @@ func (s *PricingWorkerTestSuite) TestUpdatePricesAsync() {
 		{
 			name: "several base denoms with a queued base denom",
 			baseDenoms: domain.BlockPoolMetadata{
-				DenomLiquidityMap: domain.DenomLiquidityMap{
+				UpdatedDenoms: map[string]struct{}{
 					UOSMO: {},
 					USDC:  {},
 				},
@@ -120,10 +119,10 @@ func (s *PricingWorkerTestSuite) TestUpdatePricesAsync() {
 			s.Require().False(didTimeout)
 
 			// Ensure that the correct number of base denoms are set
-			s.Require().Equal(len(tc.baseDenoms.DenomLiquidityMap), len(mockPricingUpdateListener.PricesBaseQuteDenomMap))
+			s.Require().Equal(len(tc.baseDenoms.UpdatedDenoms), len(mockPricingUpdateListener.PricesBaseQuteDenomMap))
 
 			// Ensure that non-zero prices are set for each base denom
-			s.ValidatePrices(tc.baseDenoms.DenomLiquidityMap, defaultQuoteDenom, mockPricingUpdateListener.PricesBaseQuteDenomMap)
+			s.ValidatePrices(tc.baseDenoms.UpdatedDenoms, defaultQuoteDenom, mockPricingUpdateListener.PricesBaseQuteDenomMap)
 		})
 	}
 }
@@ -166,10 +165,10 @@ func (s *PricingWorkerTestSuite) TestGetPrices_Chain_FindUnsupportedTokens() {
 
 	// Populate base denoms with all possible chain denoms
 	baseDenoms := domain.BlockPoolMetadata{
-		DenomLiquidityMap: domain.DenomLiquidityMap{},
+		UpdatedDenoms: map[string]struct{}{},
 	}
 	for chainDenom := range tokenMetadata {
-		baseDenoms.DenomLiquidityMap[chainDenom] = domain.DenomLiquidityData{}
+		baseDenoms.UpdatedDenoms[chainDenom] = struct{}{}
 	}
 
 	// Test for empty base denoms
@@ -225,7 +224,7 @@ func (s *PricingWorkerTestSuite) TestGetPrices_Chain_FindUnsupportedTokens() {
 	s.Require().Equal(20, zeroPriceCounter)
 }
 
-func (s *PricingWorkerTestSuite) ValidatePrices(initialDenoms domain.DenomLiquidityMap, expectedQuoteDenom string, prices map[string]map[string]osmomath.BigDec) {
+func (s *PricingWorkerTestSuite) ValidatePrices(initialDenoms map[string]struct{}, expectedQuoteDenom string, prices map[string]map[string]osmomath.BigDec) {
 	for baseDenom := range initialDenoms {
 		quoteMap, ok := prices[baseDenom]
 		s.Require().True(ok)