+++ CHANGELOG.md
@@ -37,6 +37,15 @@ Ref: https://keepachangelog.com/en/1.0.0/
 
 ## Unreleased
 
+- [Dynamic Min Liquidity] new ingester methods for acquiring necessary metadata.
+
+## v0.19.1
+
+- Record errors with OpenTelemetry.
+
+## v0.19.0
+
+- Ingester data transformations for dynamic min liquidity.
 - Optimize dynamic splits by caching repeated calls within callback.
 - Remove queue from pricing worker
 - Return single route quote if split quote fails